fix: allow to display request_reviewres name and refactoring api dir
